[refactoring] incapsulate demo logic
