+++ src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './pages'
 export * from './helpers'
 export * from './faker'
+export * from './demoContent'
